+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -8,10 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
  - Added controls for linear fade at the boundary of density volumes
+- Added a new TerrainLit shader that supports rendering of Unity terrains.
 
 ### Fixed
 - Stencil test during decals normal buffer update is now properly applied
 - Fixed an issue where sometimes the deferred shadow texture would not be valid, causing wrong rendering.
+
+
+### Changed
+- Changed the way depth & color pyramids are built to be faster and better quality, thus improving the look of distortion and refraction.
 
 ## [3.3.0-preview]